% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{d4_create}
\alias{d4_create}
\title{Create d4 file(s) from inputs}
\usage{
d4_create(input, output = NULL, ..., d4tools = NULL)
}
\arguments{
\item{input}{path to BigWig, BAM, or CRAM files to convert to d4.}

\item{output}{path to save d4 output. If unset, the file is named identically
and in the same path as the input file, but with the .d4 extension.}

\item{...}{see additional arguments table below for valid arguments}

\item{d4tools}{path to d4tools (or use options("d4tools" = "path/to/d4tools"))}
}
\value{
path to new d4 object
}
\description{
This function will create a new d4 file for each file passed to \code{input}.
Additional command arguments can be passed to \code{...} to configure
}
\details{
Note there is no type checking on inputs, so using \code{input}s with
heterogeneous file types may produce unintended consequences when using
file-type specific arguments in \code{...}.
}
\section{additional arguments}{
\tabular{lll}{
   name \tab data type \tab description \cr
   deflate \tab \code{logical(1)} \tab Enable deflate compression \cr
   dict_auto \tab \code{logical(1)} \tab Automatically determine dictionary type by random sampling \cr
   dump_dict \tab \code{logical(1)} \tab do not profile BAM file, only dump dictionary \cr
   sparse \tab \code{logical(1)} \tab Sparse mode, same as \verb{deflate = TRUE, dict_range = "0-1"}. Enables secondary table compression and disable primary table \cr
   deflate_level \tab \code{integer(1)} \tab Configure the deflate algorithm, default 5 \cr
   dict_file \tab \code{character(1)} \tab Path to a file defining the values of the dictionary \cr
   dict_range \tab \code{character(1)} \tab Dictionary specification as a range: "a-b" encoding values from A to B (exclusive) \cr
   filter \tab \code{character(1)} \tab A regex matching the genome name should present in the output file (note: use a shell regex, not R-style regexes) \cr
   genome \tab \code{character(1)} \tab genome description file (Used by BED inputs) \cr
   mapping_qual \tab \code{integer(1)} \tab minimal mapping quality (only valid with CRAM/BAM inputs) \cr
   ref \tab \code{character(1)} \tab path to reference genome file (only used with CRAM inputs) \cr
   threads \tab \code{integer(1)} \tab number of threads to use for encoding \cr
}
}

\examples{
\dontrun{
# Will create "example.d4"
d4_create("example.fq")

# Explicit naming is done by passing a name to "output"
d4_create("example.fq", "my_new_file.d4")
}
}
