% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{d4_hist}
\alias{d4_hist}
\title{Collect histogram of d4 file}
\usage{
d4_hist(file, regions, threads = 1, d4utils = NULL)
}
\arguments{
\item{file}{d4 file path}

\item{regions}{regions in "chr:start-end" format, or GRanges}

\item{threads}{number of threads to use}

\item{d4utils}{path to d4utils (or use options("d4utils" = "path/to/d4utils"))}
}
\value{
data.frame
}
\description{
TODO: make regions optional (use cmdfun to parse?)
}
