% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{d4_install_cargo}
\alias{d4_install_cargo}
\title{Build & Install d4utils Using Cargo}
\usage{
d4_install_cargo(..., cargo = NULL)
}
\arguments{
\item{...}{any valid cargo install flags passed as R function arguments. Omit
leading dashes, and replace dashes between words with underscores (_).}

\item{cargo}{path to cargo if installed at a nonstandard location. Default: "~/.cargo/bin/cargo".}
}
\value{

}
\description{
A convenience function for building & installing d4utils from crates.io using
cargo. Requires a Rust install and `gcc`.
}
\examples{
\dontrun{
d4_install_cargo()
}
}
