% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{d4_is_installed}
\alias{d4_is_installed}
\title{Return boolean if d4tools is installed}
\usage{
d4_is_installed(d4tools = NULL)
}
\arguments{
\item{d4tools}{path to d4tools (or use options("d4tools" = "path/to/d4tools"))}
}
\value{
bool
}
\description{
Return boolean if d4tools is installed
}
\examples{
d4_is_installed()

}
