% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{d4_stat}
\alias{d4_stat}
\title{Compute summary statistics within target regions}
\usage{
d4_stat(file, regions, method = "mean", threads = 1, d4tools = NULL)
}
\arguments{
\item{file}{path to d4 file}

\item{regions}{positions ("chr:start-end", or GRanges object). Coordiates should}

\item{method}{statistic to calculate. One of "mean" or "median".}

\item{threads}{number of threads to use}

\item{d4tools}{path to d4tools (or use options("d4tools" = "path/to/d4tools"))}
}
\value{

}
\description{
Compute summary statistics within target regions
}
