% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{d4_stat_hist_to_df}
\alias{d4_stat_hist_to_df}
\title{Title}
\usage{
d4_stat_hist_to_df(lines)
}
\arguments{
\item{lines}{}
}
\value{
data.frame("count", "n")
}
\description{
TODO: figure out how to deal with <0. It would be nice for this to be an
integer column
}
