% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{d4_view}
\alias{d4_view}
\title{View signal within regions}
\usage{
d4_view(file, regions, d4tools = NULL)
}
\arguments{
\item{file}{path to d4 file}

\item{regions}{positions ("chr:start-end", or GRanges object). Coordinates should
be 1 indexed (GRanges format), not 0 indexed (bed format).}

\item{d4tools}{path to d4tools (or use options("d4tools" = "path/to/d4tools"))}
}
\value{
GRanges object with `score` column
}
\description{
View signal within regions
}
