% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{d4_view_genome}
\alias{d4_view_genome}
\title{View genome info as Seqinfo}
\usage{
d4_view_genome(file, d4utils = NULL)
}
\arguments{
\item{file}{d4 file}

\item{d4utils}{path to d4utils (or use options("d4utils" = "path/to/d4utils"))}
}
\value{
`Seqinfo` object.
}
\description{
View genome info as Seqinfo
}
