% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{d4_install_check}
\alias{d4_install_check}
\title{Check that d4utils is installed}
\usage{
d4_install_check(d4utils = NULL)
}
\arguments{
\item{d4utils}{path to d4utils (or use options("d4utils" = "path/to/d4utils"))}
}
\value{
Prints green check if installed, red X if not detected
}
\description{
Check that d4utils is installed
}
\examples{
d4_install_check()
}
