% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{d4_is_installed}
\alias{d4_is_installed}
\title{Return boolean if d4utils is installed}
\usage{
d4_is_installed(d4utils = NULL)
}
\arguments{
\item{d4utils}{path to d4utils (or use options("d4utils" = "path/to/d4utils"))}
}
\value{
bool
}
\description{
Return boolean if d4utils is installed
}
