% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{d4_stat}
\alias{d4_stat}
\title{Compute summary statistics within target regions}
\usage{
d4_stat(
  file,
  regions,
  method = c("mean", "median"),
  threads = 1,
  d4utils = NULL
)
}
\arguments{
\item{file}{path to d4 file}

\item{regions}{positions ("chr:start-end", or GRanges object). Coordiates should}

\item{method}{statistic to calculate. One of "mean" or "median".}

\item{threads}{number of threads to use}

\item{d4utils}{path to d4utils (or use options("d4utils" = "path/to/d4utils"))}
}
\value{

}
\description{
Compute summary statistics within target regions
}
