% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{d4_view}
\alias{d4_view}
\title{View signal within regions}
\usage{
d4_view(file, regions, d4utils = NULL)
}
\arguments{
\item{file}{path do d4 file}

\item{regions}{positions ("chr:start-end", or GRanges object). Coordiates should
be 1 indexed (GRanges format), not 0 indexed (bed format).}

\item{d4utils}{path to d4utils (or use options("d4utils" = "path/to/d4utils"))}
}
\value{
GRanges object with `counts` column
}
\description{
View signal within regions
}
